requirements.txt
fastapi
pydantic
uvicorn[standard]
python-magic
google
google-cloud-firestore
firebase-admin